fix: batch-remove improper static_cast<int>(x.size()) sign-compare patterns
- Remove 6435+ occurrences of static_cast<int>(x.size()) in size/size
  comparisons across 1107 src/ files where both sides were unsigned
- Fix pre-existing GCC compile errors in ts_operator_diagnostics.cpp and
  adaptive_query_compiler.cpp (invalid = token inside cast expression)
- Fix malformed cast in multi_gpu_vector_index.cpp:306
- Fix ruleCount() wrong return type in prompt_policy.cpp
- Fix sign-compare for-loop variable in tensor_aware_query_optimizer.cpp
- Restore static_cast<int> where narrowing is semantically significant
  (OpenSSL EVP_CTRL_GCM_SET_TAG, heap slot index, node_idx, M/N bounds,
   numClasses(), LDAP idle pool count)

+++ src/acceleration/ai_hardware_dispatcher.cpp
@@ -963,7 +963,7 @@ AiInferenceResult AiHardwareDispatcher::dispatchOnnxRuntime(AiInferenceRequest &
         // Input tensor
         OrtValue *input_tensor = nullptr;
         ort->CreateTensorWithDataAsOrtValue(mem_info, const_cast<float *>(req.input_data),
-                                            req.input_elements * sizeof(float), shape.data(),static_cast<int>(shape.size()),
+                                            req.input_elements * sizeof(float), shape.data(),shape.size(),
                                             ONNX_TENSOR_ELEMENT_DATA_TYPE_FLOAT, &input_tensor);
 
         // Query input/output names
